Adds final test case and resends the Exception.
The reason I have chosen to throw an exception
instead of catching the exception and printing
a message is twofold.

Firstly, we cannot test the use of an exit()
or die() that would be required to catch the
error and present a message alerting the user
of the issue.

Secondly, by handling the error, it prevents
consumming libraries from easily detecting an
error has been thown and handling it gracefully.

Both of these reasons have lead me to determining
that Exceptions should be handled in code by
consuming libraries. I'll document this on the
README so users can handle the exceptions if they
wish. I'll also have to use this in Acquia Cli
and Acquia Logstream. I don't catch errors in
these at all right now so it would be an
enhancement!